requirements.txt
pandas>=1.4.2
seaborn>=0.11.2
torchvision>=0.12.0
torch>=1.11.0
scipy>=1.3.3
pygame>=2.1.0
matplotlib>=3.1.2
torchbooster>=0.0.1
numpy>=1.17.4
tqdm>=4.64.0
dataclasses>=0.8
pyserial>=3.5
sklearn>=1.0.2
torchinfo>=1.6.5